TQ-aware prompt cache + audit fixes: bn_alloc, size_t casts, warning cleanup
Prompt cache now stores TQ-compressed KV data natively (8.7x smaller entries,
10x faster restore). Split kv_bytes into key_cache_bytes/val_cache_bytes,
added kv_tq_bits format validation to prevent FP32/TQ cache mismatches.

Audit fixes: replace malloc/free with bn_alloc in moe batch path, add
(size_t) casts to prevent int overflow in array indexing (moe.c, model.c,
prompt_cache.c), eliminate compiler warnings (unused vars on non-AVX2),
replace atoi with strtol in bench.

+++ src/model.c
@@ -1034,8 +1034,8 @@ int bn_model_alloc_session_buffers(const BnConfig *c, const BnWeights *w,
         if (bn_tq_init(&tq_tmp, c->head_size, c->kv_tq_bits, 0) == 0) {
             int key_bytes = bn_tq_key_bytes(&tq_tmp);
             int val_bytes = bn_tq_value_bytes(&tq_tmp);
-            size_t tq_key_total = (size_t)n_attn_layers * c->seq_len * c->n_kv_heads * key_bytes;
-            size_t tq_val_total = (size_t)n_attn_layers * c->seq_len * c->n_kv_heads * val_bytes;
+            size_t tq_key_total = (size_t)n_attn_layers * (size_t)c->seq_len * (size_t)c->n_kv_heads * (size_t)key_bytes;
+            size_t tq_val_total = (size_t)n_attn_layers * (size_t)c->seq_len * (size_t)c->n_kv_heads * (size_t)val_bytes;
             s->key_cache_tq   = (uint8_t *)sh_arena_calloc(arena, tq_key_total, 1);
             s->value_cache_tq = (uint8_t *)sh_arena_calloc(arena, tq_val_total, 1);
             s->q_rotated      = (float *)sh_arena_calloc(arena, c->n_heads * c->head_size, sizeof(float));
